requirements.txt
# Backend
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Frontend
streamlit>=1.28.0
pyvis>=0.3.2
plotly>=5.15.0

# Data Processing
pandas>=2.0.0
requests>=2.31.0

